move example files to /example dir
