+++ README.md
@@ -105,7 +105,7 @@ server.register([require('hapi-error'), require('vision')], function (err) {
 module.exports = server;
 ```
 
-> See: [/server_example.js]() for simple example
+> See: [/example/server_example.js](https://github.com/dwyl/hapi-error/blob/master/example/server_example.js) for simple example
 
 ### 3. Ensure that you have a View called `error_template`
 
@@ -114,7 +114,15 @@ which allows you to use Handlebars, Jade, React, etc. for your templates.
 
 Your `error_template.html` (*or `error_template.ext` `error_template.jsx`*) should make use of the 3 variables it will be passed:
 
++ `errorTitle` - *the error tile generated by Hapi*
++ `statusCode` - *HTTP statusCode sent to the client *e.g: `404`* (*not found*)
++ `errorMessage` - the *human-friendly error message*
 
+> for an example see: [`/example/error_template.html`](https://github.com/dwyl/hapi-error/blob/master/example/error_template.html)
+
+That's it!
+
+---
 
 ### Implementation Detail:
